fixed decoding accuracy
The calculation of decoding accuracy as proportion correct (for classification) or correlation coefficient (for reconstruction) was assuming that all trials are being classified or reconstructed. However, this is not necessarily the case, e.g. when classifying between 2 out of 4 conditions or when reconstructing a parametric modulator belonging to just 1 condition. As a consequence, decoding accuracy was underestimated. The new code is more general in the way described above.