edit trial-wise regression
The default value of the "mode" parameter of "ITEM_est_1st_lvl.m" was changed from 'DCT' to 'KXY'. This means that, when not specified, temporal filter regressors are not included into the design matrix X_t (which means they will also affect U) anymore, but data and design are filtered beforehand (like in SPM) using the residual-forming matrix approach (like in SPM). It seems that some people only use the trial-wise regression function of this toolbox and then use their own customized analysis to analyze the trial-wise parameter estimates. In this situation, 'KXY' is the more appropriate choice, because trial-wise parameter estimates are not contaminated with filter regressor variance (which would have to be accounted for via U) in this case. Note that this change does not affect the analyses in the "ITEM-paper" repository, because the "mode" parameter is explicitly specified as 'DCT' in these scripts.